feature/era-glonass-integration-r2:
moved type-families generic helpers from
__carma-era-glonass-integration__ package to __carma-utils__
